+++ src/Futhark/CodeGen/ImpGen/Kernels/SegHist.hs
@@ -199,7 +199,7 @@ data Passage = MustBeSinglePass | MayBeMultiPass deriving (Eq, Ord)
 
 bodyPassage :: KernelBody KernelsMem -> Passage
 bodyPassage kbody
-  | mempty == consumedInKernelBody (aliasAnalyseKernelBody kbody) =
+  | mempty == consumedInKernelBody (aliasAnalyseKernelBody mempty kbody) =
     MayBeMultiPass
   | otherwise =
     MustBeSinglePass